backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy>=2.0.27
pymysql==1.1.0
cryptography==41.0.7